feat: Add configs to discount badge
