+++ packages/core/src/components/ui/ProductShelf/Overrides.tsx
@@ -3,8 +3,8 @@ import ProductCard from 'src/components/product/ProductCard'
 import ProductShelfCustomizations from 'src/customizations/components/overrides/ProductShelf'
 
 const Components = {
-    ProductCard,
-    ...ProductShelfCustomizations.components
+  ProductCard,
+  ...ProductShelfCustomizations.components,
 }
 
 export { Components }
