+++ packages/core/src/components/ui/ProductShelf/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ProductShelf'
+export type { ProductShelfProps } from './ProductShelf'
